🔥 Clean structure a bit and add comments
