+++ crates/fuzz/src/lib.rs
@@ -67,7 +67,17 @@ pub mod fuzzing {
     pub use tokio;
 
     #[cfg(feature = "token")]
-    pub use super::trident::token2022::*;
+    pub use super::trident::AccountExtension;
+    #[cfg(feature = "token")]
+    pub use super::trident::MintExtension;
+    #[cfg(feature = "token")]
+    pub use super::trident::MintExtensionData;
+    #[cfg(feature = "token")]
+    pub use super::trident::MintWithExtensions;
+    #[cfg(feature = "token")]
+    pub use super::trident::TokenAccountExtensionData;
+    #[cfg(feature = "token")]
+    pub use super::trident::TokenAccountWithExtensions;
 
     extern "C" {
         pub fn __llvm_profile_set_filename(filename: *const i8);
