+++ crates/fuzz/src/trident/client.rs
@@ -5,6 +5,7 @@ use solana_sdk::transaction::TransactionError;
 use trident_svm::prelude::solana_svm::transaction_processing_result::TransactionProcessingResult;
 use trident_svm::processor::InstructionError;
 
+use crate::trident::transaction_result::TransactionResult;
 use crate::trident::Trident;
 
 use solana_sdk::account::AccountSharedData;
@@ -19,54 +20,26 @@ use solana_sdk::sysvar::Sysvar;
 use trident_svm::types::trident_entrypoint::TridentEntrypoint;
 use trident_svm::types::trident_program::TridentProgram;
 
-pub struct TransactionResult {
-    transaction_result: solana_sdk::transaction::Result<()>,
-    transaction_logs: Vec<String>,
-}
-
-impl TransactionResult {
-    fn new(
-        transaction_result: solana_sdk::transaction::Result<()>,
-        transaction_logs: Vec<String>,
-    ) -> Self {
-        Self {
-            transaction_result,
-            transaction_logs,
-        }
-    }
-
-    pub fn is_success(&self) -> bool {
-        self.transaction_result.is_ok()
-    }
-
-    pub fn is_error(&self) -> bool {
-        self.transaction_result.is_err()
-    }
-
-    pub fn logs(&self) -> &[String] {
-        &self.transaction_logs
-    }
-    pub fn get_result(&self) -> &solana_sdk::transaction::Result<()> {
-        &self.transaction_result
-    }
-    pub fn get_custom_error_code(&self) -> Option<u32> {
-        self.transaction_result
-            .as_ref()
-            .err()
-            .and_then(|result| match result {
-                TransactionError::InstructionError(
-                    _error_code,
-                    InstructionError::Custom(error_code),
-                ) => Some(*error_code),
-                _ => None,
-            })
-    }
-    pub fn is_custom_error_with_code(&self, error_code: u32) -> bool {
-        self.get_custom_error_code() == Some(error_code)
-    }
-}
-
 impl Trident {
+    /// Processes a transaction containing one or more instructions
+    ///
+    /// This method executes the provided instructions as a single transaction and returns
+    /// the result including success/failure status and transaction logs. It also handles
+    /// fuzzing metrics collection when enabled via environment variables.
+    ///
+    /// # Arguments
+    /// * `instructions` - A slice of instructions to execute in the transaction
+    /// * `transaction_name` - A descriptive name for the transaction (used in metrics)
+    ///
+    /// # Returns
+    /// A `TransactionResult` containing the execution result and logs
+    ///
+    /// # Example
+    /// ```rust,ignore
+    /// let instructions = vec![system_instruction::transfer(&from, &to, 1000)];
+    /// let result = trident.process_transaction(&instructions, "Transfer SOL");
+    /// assert!(result.is_success());
+    /// ```
     pub fn process_transaction(
         &mut self,
         instructions: &[Instruction],
@@ -93,15 +66,36 @@ impl Trident {
         self.handle_tx_result(tx_result, transaction_name, instructions)
     }
 
+    /// Deploys an entrypoint program to the SVM runtime
+    ///
+    /// This method is only available when the "syscall-v2" feature is enabled.
+    /// It deploys a program that serves as an entrypoint for other programs.
+    ///
+    /// # Arguments
+    /// * `_program` - The entrypoint program to deploy
     #[cfg(feature = "syscall-v2")]
     pub fn deploy_entrypoint(&mut self, _program: TridentEntrypoint) {
         self.client.deploy_entrypoint_program(&_program);
     }
 
+    /// Deploys a binary program to the SVM runtime
+    ///
+    /// This method deploys a compiled Solana program (BPF/SBF) to the runtime,
+    /// making it available for instruction execution.
+    ///
+    /// # Arguments
+    /// * `program` - The compiled program to deploy
     pub fn deploy_program(&mut self, program: TridentProgram) {
         self.client.deploy_binary_program(&program);
     }
 
+    /// Warps the blockchain clock to a specific epoch
+    ///
+    /// This method updates the system clock sysvar to simulate time progression
+    /// to the specified epoch, useful for testing time-dependent program logic.
+    ///
+    /// # Arguments
+    /// * `warp_epoch` - The target epoch to warp to
     pub fn warp_to_epoch(&mut self, warp_epoch: u64) {
         let mut clock = self.get_sysvar::<Clock>();
 
@@ -109,12 +103,26 @@ impl Trident {
         self.client.set_sysvar(&clock);
     }
 
+    /// Warps the blockchain clock to a specific slot
+    ///
+    /// This method updates the system clock sysvar to simulate progression
+    /// to the specified slot number.
+    ///
+    /// # Arguments
+    /// * `warp_slot` - The target slot number to warp to
     pub fn warp_to_slot(&mut self, warp_slot: u64) {
         let mut clock = self.get_sysvar::<Clock>();
 
         clock.slot = warp_slot;
         self.client.set_sysvar(&clock);
     }
+    /// Warps the blockchain clock to a specific Unix timestamp
+    ///
+    /// This method updates the system clock sysvar to simulate time progression
+    /// to the specified Unix timestamp.
+    ///
+    /// # Arguments
+    /// * `warp_timestamp` - The target Unix timestamp to warp to
     pub fn warp_to_timestamp(&mut self, warp_timestamp: i64) {
         let mut clock = self.get_sysvar::<Clock>();
 
@@ -122,6 +130,13 @@ impl Trident {
         self.client.set_sysvar(&clock);
     }
 
+    /// Advances the blockchain clock by a specified number of seconds
+    ///
+    /// This method increments the current Unix timestamp by the given number
+    /// of seconds, useful for testing time-based program behavior.
+    ///
+    /// # Arguments
+    /// * `seconds` - The number of seconds to advance the clock
     pub fn forward_in_time(&mut self, seconds: i64) {
         let mut clock = self.get_sysvar::<Clock>();
 
@@ -129,17 +144,53 @@ impl Trident {
         self.client.set_sysvar(&clock);
     }
 
+    /// Sets a custom account state at the specified address
+    ///
+    /// This method allows you to manually set account data, lamports, and owner
+    /// for any public key, useful for setting up test scenarios.
+    ///
+    /// # Arguments
+    /// * `address` - The public key where the account should be stored
+    /// * `account` - The account data to set
     pub fn set_account_custom(&mut self, address: &Pubkey, account: &AccountSharedData) {
         self.client.set_account(address, account, false);
     }
 
+    /// Returns the default payer keypair for transactions
+    ///
+    /// This keypair is used to pay transaction fees and sign transactions
+    /// when no other payer is specified.
+    ///
+    /// # Returns
+    /// The default payer keypair
     pub fn payer(&self) -> solana_sdk::signature::Keypair {
         self.client.get_payer()
     }
 
+    /// Retrieves account data for the specified public key
+    ///
+    /// Returns the account data including lamports, owner, and data bytes.
+    /// If the account doesn't exist, returns a default empty account.
+    ///
+    /// # Arguments
+    /// * `key` - The public key of the account to retrieve
+    ///
+    /// # Returns
+    /// The account data or a default account if not found
     pub fn get_account(&mut self, key: &Pubkey) -> AccountSharedData {
         trident_svm::trident_svm::TridentSVM::get_account(&self.client, key).unwrap_or_default()
     }
+    /// Retrieves and deserializes account data as a specific type
+    ///
+    /// This method fetches account data and attempts to deserialize it using Borsh,
+    /// skipping the specified discriminator bytes at the beginning.
+    ///
+    /// # Arguments
+    /// * `key` - The public key of the account to retrieve
+    /// * `discriminator_size` - Number of bytes to skip before deserializing
+    ///
+    /// # Returns
+    /// Some(T) if deserialization succeeds, None otherwise
     pub fn get_account_with_type<T: BorshDeserialize>(
         &mut self,
         key: &Pubkey,
@@ -155,10 +206,20 @@ impl Trident {
         }
     }
 
+    /// Gets the current Unix timestamp from the blockchain clock
+    ///
+    /// Returns the current timestamp as stored in the Clock sysvar.
+    ///
+    /// # Returns
+    /// The current Unix timestamp in seconds
     pub fn get_current_timestamp(&self) -> i64 {
         self.get_sysvar::<Clock>().unix_timestamp
     }
 
+    /// Gets the last blockhash (not implemented for TridentSVM)
+    ///
+    /// # Panics
+    /// This method always panics as it's not yet implemented for TridentSVM
     pub fn get_last_blockhash(&self) -> Hash {
         panic!("Not yet implemented for TridentSVM");
     }
@@ -179,10 +240,25 @@ impl Trident {
         self.client.process_transaction_with_settle(tx)
     }
 
+    /// Retrieves a system variable (sysvar) of the specified type
+    ///
+    /// System variables contain blockchain state information like clock,
+    /// rent, and epoch schedule data.
+    ///
+    /// # Returns
+    /// The requested sysvar data
     pub fn get_sysvar<T: Sysvar>(&self) -> T {
         trident_svm::trident_svm::TridentSVM::get_sysvar::<T>(&self.client)
     }
 
+    /// Airdrops SOL to the specified address
+    ///
+    /// This method adds the specified amount of lamports to the target account.
+    /// If the account doesn't exist, it will be created with the airdropped amount.
+    ///
+    /// # Arguments
+    /// * `address` - The public key to receive the airdrop
+    /// * `amount` - The number of lamports to airdrop
     pub fn airdrop(&mut self, address: &Pubkey, amount: u64) {
         let mut account = self.get_account(address);
 
@@ -268,49 +344,5 @@ impl Trident {
             },
             Err(transaction_error) => TransactionResult::new(Err(transaction_error.clone()), vec![]),
         }
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn create_account(
-        &mut self,
-        address: &Pubkey,
-        from: &Pubkey,
-        space: usize,
-        owner: &Pubkey,
-    ) -> Vec<Instruction> {
-        let account = self.client.get_account(address).unwrap_or_default();
-        let rent = solana_sdk::rent::Rent::default();
-        if account.lamports() > 0 {
-            let mut instructions = vec![];
-            let lamports_required = rent.minimum_balance(space);
-
-            let remaining_lamports = lamports_required.saturating_sub(account.lamports());
-
-            if remaining_lamports > 0 {
-                let transfer =
-                    solana_sdk::system_instruction::transfer(from, address, remaining_lamports);
-                instructions.push(transfer);
-            }
-
-            let allocate = solana_sdk::system_instruction::allocate(address, space as u64);
-            instructions.push(allocate);
-
-            let assign = solana_sdk::system_instruction::assign(address, owner);
-            instructions.push(assign);
-
-            instructions
-        } else {
-            let ix = solana_sdk::system_instruction::create_account(
-                from,
-                address,
-                rent.minimum_balance(space),
-                space as u64,
-                owner,
-            );
-            vec![ix]
-        }
-    }
-    pub fn transfer(&mut self, from: &Pubkey, to: &Pubkey, amount: u64) -> Instruction {
-        solana_sdk::system_instruction::transfer(from, to, amount)
     }
 }
