+++ crates/fuzz/src/trident/flow_executor.rs
@@ -23,18 +23,52 @@ mod config {
     pub const ENV_FUZZING_METRICS: &str = "FUZZING_METRICS";
 }
 
+/// Trait for executing fuzzing flows in the Trident framework
+///
+/// This trait defines the interface for fuzzing executors that can run
+/// multiple iterations of randomized program interactions. Implementors
+/// should provide the core fuzzing logic while this trait handles
+/// parallelization, progress tracking, and metrics collection.
 pub trait FlowExecutor: Send + 'static + Sized {
+    /// Creates a new instance of the flow executor
     fn new() -> Self;
+
+    /// Executes a specified number of flow calls in a single iteration
+    ///
+    /// # Arguments
+    /// * `flow_calls_per_iteration` - Number of individual flow calls to execute
+    ///
+    /// # Returns
+    /// Result indicating success or a fuzzing error
     fn execute_flows(
         &mut self,
         flow_calls_per_iteration: u64,
     ) -> Result<(), crate::error::FuzzingError>;
+
+    /// Returns a mutable reference to the underlying Trident instance
     fn trident_mut(&mut self) -> &mut Trident;
+
+    /// Resets fuzz accounts to their initial state for the next iteration
     fn reset_fuzz_accounts(&mut self);
 
-    /// Generated by macro: empty or contains LLVM coverage calls
+    /// Handles LLVM coverage collection (generated by macro)
+    ///
+    /// This method is typically empty or contains LLVM coverage calls
+    /// depending on whether coverage profiling is enabled.
+    ///
+    /// # Arguments
+    /// * `current_iteration` - The current iteration number
     fn handle_llvm_coverage(&mut self, current_iteration: u64);
 
+    /// Main entry point for fuzzing execution
+    ///
+    /// This method orchestrates the entire fuzzing process, handling both
+    /// single-threaded and parallel execution based on the environment
+    /// and available system resources.
+    ///
+    /// # Arguments
+    /// * `iterations` - Total number of fuzzing iterations to run
+    /// * `flow_calls_per_iteration` - Number of flow calls per iteration
     fn fuzz(iterations: u64, flow_calls_per_iteration: u64) {
         if std::env::var(config::ENV_FUZZ_DEBUG).is_ok() {
             println!("Debug mode detected: Running single iteration with provided seed");
