+++ crates/fuzz/src/trident/metrics.rs
@@ -2,6 +2,18 @@ use crate::trident::Trident;
 use solana_sdk::pubkey::Pubkey;
 
 impl Trident {
+    /// Records a value in a histogram metric
+    ///
+    /// Histogram metrics track the distribution of values over time,
+    /// useful for measuring performance characteristics like execution times,
+    /// gas usage, or other numerical distributions.
+    ///
+    /// # Arguments
+    /// * `metric_name` - Name of the histogram metric
+    /// * `value` - The value to record in the histogram
+    ///
+    /// # Note
+    /// Metrics are only recorded when the FUZZING_METRICS environment variable is set
     pub fn record_histogram(&mut self, metric_name: &str, value: f64) {
         let metrics = std::env::var("FUZZING_METRICS");
         if metrics.is_ok() {
@@ -9,6 +21,17 @@ impl Trident {
         }
     }
 
+    /// Records a value in an accumulator metric
+    ///
+    /// Accumulator metrics sum up values over time, useful for tracking
+    /// totals like total gas consumed, total tokens transferred, etc.
+    ///
+    /// # Arguments
+    /// * `metric_name` - Name of the accumulator metric
+    /// * `value` - The value to add to the accumulator
+    ///
+    /// # Note
+    /// Metrics are only recorded when the FUZZING_METRICS environment variable is set
     pub fn record_accumulator(&mut self, metric_name: &str, value: f64) {
         let metrics = std::env::var("FUZZING_METRICS");
         if metrics.is_ok() {
@@ -16,6 +39,18 @@ impl Trident {
         }
     }
 
+    /// Tracks account state for regression testing
+    ///
+    /// This method captures the current state of an account for regression
+    /// analysis, allowing you to detect when account states change unexpectedly
+    /// between fuzzing runs.
+    ///
+    /// # Arguments
+    /// * `account` - The public key of the account to track
+    /// * `account_name` - A descriptive name for the account
+    ///
+    /// # Note
+    /// Account tracking is only enabled when the FUZZING_REGRESSION environment variable is set
     pub fn track_account_regression(&mut self, account: &Pubkey, account_name: &str) {
         let regression = std::env::var("FUZZING_REGRESSION");
         if regression.is_ok() {
